feat(v2/interface): provide proper FileManager interface and merge v2 types, functions
chore: rename fileInfoList to recordList
