+++ src/fileManager.ts
@@ -17,7 +17,8 @@ import {
   Topic,
 } from '@ethersphere/bee-js';
 
-import { DownloadResource, DownloadResult } from './types/v2/download';
+import { DownloadResource, DownloadResult } from './types/download';
+import { FileManager } from './types/fileManager';
 import {
   DriveInfo,
   FileRecord,
@@ -29,9 +30,10 @@ import {
   NodeStatus,
   NodeType,
   TrashEntry,
-} from './types/v2/info';
-import { UpdateItem, UploadFilesResult, UploadItem } from './types/v2/upload';
-import { ActReferences } from './types/v2/utils';
+} from './types/info';
+import { UpdateItem, UploadFilesResult, UploadItem } from './types/upload';
+import { ActReferences } from './types/utils';
+import { assertActReferences, assertDriveInfoFromMetadata, assertReady } from './utils/asserts';
 import {
   fetchStamp,
   getFeedData,
@@ -39,7 +41,7 @@ import {
   verifyStampUsability,
   verifySupportedBeeVersions,
 } from './utils/bee';
-import { settlePromises } from './utils/common';
+import { awaitAllPromisesBounded, getRecordStatus, joinPath, settlePromises } from './utils/common';
 import {
   ADMIN_STAMP_LABEL,
   FEED_INDEX_ZERO,
@@ -53,11 +55,9 @@ import {
   ROOT_PATH,
 } from './utils/constants';
 import { generateRandomBytes } from './utils/crypto';
-import { DriveError, ErrorHandler, FileInfoError, SignerError, StampError } from './utils/errors';
+import { DriveError, ErrorHandler, FileRecordError, SignerError, StampError } from './utils/errors';
 import { FileManagerEvents } from './utils/events';
 import { Logger } from './utils/logger';
-import { assertActReferences, assertDriveInfoFromMetadata, assertReady } from './utils/v2/asserts';
-import { awaitAllPromisesBounded, getRecordStatus, joinPath } from './utils/v2/common';
 import {
   driveForkMetadata,
   fileForkMetadata,
@@ -65,15 +65,14 @@ import {
   getAllNodeEntries,
   getDriveForkPath,
   getRlevel,
-} from './utils/v2/mantaray';
-import { assertValidRelativePath, normalizePath, pathSegments, splitPath } from './utils/v2/path';
+} from './utils/mantaray';
+import { assertValidRelativePath, normalizePath, pathSegments, splitPath } from './utils/path';
 import { processDownload } from './download';
 import { EventEmitter, EventEmitterBase } from './eventEmitter';
 import { MantarayStore } from './mantarayStore';
 import { assertUploadableSource, processUpload } from './upload';
 
-// TODO: restore `implements FileManager` when all methods land
-export class FileManagerBase {
+export class FileManagerBase implements FileManager {
   private bee: Bee;
   private signer: PrivateKey;
   private signerAddress: string;
@@ -81,21 +80,29 @@ export class FileManagerBase {
   private stateFeedTopic: Topic | undefined = undefined;
   private _isInitialized: boolean = false;
   private isInitializing: boolean = false;
-  private _adminStamp: PostageBatch | undefined = undefined;
   private adminRedundancyLevel: RedundancyLevel = RedundancyLevel.OFF;
 
+  private _adminStamp: PostageBatch | undefined = undefined;
+  private readonly _driveList: DriveInfo[] = [];
+  private readonly _recordList: FileRecord[] = [];
   private readonly store: MantarayStore;
   private readonly errorHandler = ErrorHandler.getInstance();
   private readonly logger = Logger.getInstance();
 
   // --- Public member getters ---
 
-  readonly driveList: DriveInfo[] = [];
-  readonly fileInfoList: FileRecord[] = [];
   readonly emitter: EventEmitter;
 
   get adminStamp(): PostageBatch | undefined {
     return this._adminStamp;
+  }
+
+  get driveList(): readonly DriveInfo[] {
+    return this._driveList;
+  }
+
+  get recordList(): readonly FileRecord[] {
+    return this._recordList;
   }
 
   get isInitialized(): boolean {
@@ -348,7 +355,7 @@ export class FileManagerBase {
     const { driveIx, cachedDrive } = this.findDriveOrThrow(new Identifier(driveId).toString());
 
     if (!items.length) {
-      throw new FileInfoError('uploadFiles requires at least one entry');
+      throw new FileRecordError('uploadFiles requires at least one entry');
     }
 
     for (const entry of items) {
@@ -430,7 +437,7 @@ export class FileManagerBase {
 
       const folderTopic = meta[MANIFEST_METADATA_NODE_TOPIC];
       if (!folderTopic) {
-        throw new FileInfoError(`Folder fork missing topic: ${path}`);
+        throw new FileRecordError(`Folder fork missing topic: ${path}`);
       }
 
       // Folder manifest reads always probe the feed head. A folder is a container and carries, no stored version
@@ -493,7 +500,7 @@ export class FileManagerBase {
 
         const parentHost = hostMap.get(planned.parentPath);
         if (!parentHost) {
-          throw new FileInfoError(`Internal error: parent folder not resolved for path: ${planned.fullPath}`);
+          throw new FileRecordError(`Internal error: parent folder not resolved for path: ${planned.fullPath}`);
         }
 
         const { topic, version } = await getTopicAndVersion(this.bee, owner, undefined, undefined, requestOptions);
@@ -580,7 +587,7 @@ export class FileManagerBase {
 
     const noMeta = !changes.customMetadata || Object.keys(changes.customMetadata).length === 0;
     if (noMeta && !changes.item) {
-      throw new FileInfoError('Neither a file/path nor customMetadata is provided');
+      throw new FileRecordError('Neither a file/path nor customMetadata is provided');
     }
 
     const owner = this.signerAddress;
@@ -594,7 +601,7 @@ export class FileManagerBase {
     );
 
     if (cached.driveId !== cachedDrive.id) {
-      throw new FileInfoError(`Record ${record.topic.slice(0, 6)} does not belong to drive "${cachedDrive.name}"`);
+      throw new FileRecordError(`Record ${record.topic.slice(0, 6)} does not belong to drive "${cachedDrive.name}"`);
     }
 
     // A cached record already holds the authoritative absolute path; keep it.
@@ -704,7 +711,7 @@ export class FileManagerBase {
   ): Promise<FileRecord> {
     assertReady(this.publisher, this.isInitialized, this.stateFeedTopic);
 
-    const localHead = this.fileInfoList.find((f) => f.topic === fr.topic);
+    const localHead = this.recordList.find((f) => f.topic === fr.topic);
 
     if (localHead && localHead.version && version) {
       const requested = new FeedIndex(version);
@@ -718,7 +725,7 @@ export class FileManagerBase {
     const index = version !== undefined ? new FeedIndex(version).toBigInt() : undefined;
     const feedData = await getFeedData(this.bee, topic, fr.owner, index, requestOptions);
     if (feedData.feedIndex.equals(FeedIndex.MINUS_ONE)) {
-      throw new FileInfoError(`File feed not found for topic: ${fr.topic.slice(0, 6)}`);
+      throw new FileRecordError(`File feed not found for topic: ${fr.topic.slice(0, 6)}`);
     }
 
     return this.store.getRecord(topic.toString(), fr.actPublisher, feedData, requestOptions);
@@ -736,21 +743,21 @@ export class FileManagerBase {
       requestOptions,
     );
     if (feedIndex.equals(FeedIndex.MINUS_ONE.toString())) {
-      throw new FileInfoError('Record feed not found');
+      throw new FileRecordError('Record feed not found');
     }
 
     if (!versionToRestore.version) {
-      throw new FileInfoError('Restore version has to be defined');
+      throw new FileRecordError('Restore version has to be defined');
     }
 
     const versionToRestoreIndex = new FeedIndex(versionToRestore.version);
     if (feedIndex.equals(versionToRestoreIndex)) {
-      throw new FileInfoError(
+      throw new FileRecordError(
         `Head Slot cannot be restored. Please select a version lesser than: ${versionToRestore.version}`,
       );
     }
 
-    const cached = this.fileInfoList.find((f) => f.topic === versionToRestore.topic);
+    const cached = this.recordList.find((f) => f.topic === versionToRestore.topic);
 
     const newVersion = feedIndexNext.toString();
     const restored: FileRecord = {
@@ -875,7 +882,7 @@ export class FileManagerBase {
         }),
         MAX_CONCURRENT_FEED_FETCHES,
         (record) => {
-          if (!this.fileInfoList.some((f) => f.topic === record.topic)) this.fileInfoList.push(record);
+          if (!this.recordList.some((f) => f.topic === record.topic)) this._recordList.push(record);
           results.push(record);
         },
         (reason, ix) => {
@@ -960,7 +967,7 @@ export class FileManagerBase {
 
     const normalized = normalizePath(path);
     const prefix = normalized ? normalized + '/' : '';
-    const files = this.fileInfoList.filter((f) => f.driveId === driveId.toString() && f.path.startsWith(prefix));
+    const files = this.recordList.filter((f) => f.driveId === driveId.toString() && f.path.startsWith(prefix));
 
     return this.downloadFiles(files, options, requestOptions);
   }
@@ -1020,7 +1027,7 @@ export class FileManagerBase {
 
     const movedTopic = forkMetadata[MANIFEST_METADATA_NODE_TOPIC];
     if (movedTopic && getRecordStatus(cachedSource, movedTopic) === NodeStatus.Trashed) {
-      throw new FileInfoError('Cannot move a trashed file/folder; recover it first');
+      throw new FileRecordError('Cannot move a trashed file/folder; recover it first');
     }
 
     const { host: tgtParentHost, folder: tgtParentFolder } = await this.store.resolveHost(
@@ -1038,7 +1045,7 @@ export class FileManagerBase {
     if (isFile) {
       const fileTopic = forkMetadata[MANIFEST_METADATA_FILE_TOPIC];
       if (!fileTopic) {
-        throw new FileInfoError(`Fork at ${fromPath} has no file topic — cannot move`);
+        throw new FileRecordError(`Fork at ${fromPath} has no file topic — cannot move`);
       }
 
       const { record } = await this.loadRecord(fileTopic, this.signerAddress, publisher, undefined, requestOptions);
@@ -1081,7 +1088,7 @@ export class FileManagerBase {
       // reset the in-memory cache
       const fromPrefix = fromPath + '/';
       const toPrefix = toPath + '/';
-      for (const f of this.fileInfoList) {
+      for (const f of this.recordList) {
         if (f.driveId === sourceDriveId.toString() && f.path.startsWith(fromPrefix)) {
           f.path = toPrefix + f.path.substring(fromPrefix.length);
           if (isCrossDrive) {
@@ -1133,7 +1140,7 @@ export class FileManagerBase {
 
     const fork = parentNode.find(name);
     if (!fork) {
-      throw new FileInfoError(`Path not found: ${path}`);
+      throw new FileRecordError(`Path not found: ${path}`);
     }
 
     const meta = fork.metadata ?? {};
@@ -1153,10 +1160,10 @@ export class FileManagerBase {
       }
 
       const prefix = path.endsWith('/') ? path : path + '/';
-      for (let i = this.fileInfoList.length - 1; i >= 0; --i) {
-        const f = this.fileInfoList[i];
+      for (let i = this.recordList.length - 1; i >= 0; --i) {
+        const f = this.recordList[i];
         if (f.driveId === cachedDrive.id && f.path.startsWith(prefix)) {
-          this.fileInfoList.splice(i, 1);
+          this._recordList.splice(i, 1);
         }
       }
 
@@ -1166,11 +1173,11 @@ export class FileManagerBase {
       return;
     }
 
-    const fiIndex = this.fileInfoList.findIndex((f) => f.driveId === cachedDrive.id && f.path === path);
+    const fiIndex = this.recordList.findIndex((f) => f.driveId === cachedDrive.id && f.path === path);
 
-    const forgotten = fiIndex !== -1 ? this.fileInfoList[fiIndex] : undefined;
+    const forgotten = fiIndex !== -1 ? this.recordList[fiIndex] : undefined;
     if (fiIndex !== -1) {
-      this.fileInfoList.splice(fiIndex, 1);
+      this._recordList.splice(fiIndex, 1);
     }
 
     await this.pruneTrashOverlay(driveIx, (n) => n.topic === nodeTopic, requestOptions);
@@ -1377,7 +1384,7 @@ export class FileManagerBase {
     adminMantaray.addFork(getDriveForkPath(newDrive.id), new Reference(newDrive.topic), driveForkMetadata(newDrive));
     await this.store.saveMantarayNode(adminMantaray, adminHost, requestOptions);
 
-    this.driveList.push(newDrive);
+    this._driveList.push(newDrive);
     this.emitter.emit(FileManagerEvents.DRIVE_CREATED, { driveInfo: newDrive });
 
     return newDrive;
@@ -1402,7 +1409,7 @@ export class FileManagerBase {
     }
 
     if (reset) {
-      this.driveList.length = 0;
+      this._driveList.length = 0;
       this.store.clear();
     }
 
@@ -1438,7 +1445,7 @@ export class FileManagerBase {
     if (!this.publisher) {
       throw new SignerError('Publisher not found');
     }
-    // TODO: possible performance improvement: store version next to stateFeedTopic
+
     const { payload, feedIndex, feedIndexNext } = await getFeedData(
       this.bee,
       this.stateFeedTopic,
@@ -1506,7 +1513,7 @@ export class FileManagerBase {
       }),
       (driveInfo) => {
         if (driveInfo) {
-          this.driveList.push(driveInfo);
+          this._driveList.push(driveInfo);
         }
       },
       (reason) =>
@@ -1566,12 +1573,12 @@ export class FileManagerBase {
     adminMantaray.removeFork(getDriveForkPath(driveInfo.id));
     await this.store.saveMantarayNode(adminMantaray, adminHost, requestOptions);
 
-    this.driveList.splice(driveIndex, 1);
+    this._driveList.splice(driveIndex, 1);
     this.store.evict(driveInfo.topic);
 
-    for (let i = this.fileInfoList.length - 1; i >= 0; --i) {
-      if (this.fileInfoList[i].driveId === driveInfo.id) {
-        this.fileInfoList.splice(i, 1);
+    for (let i = this.recordList.length - 1; i >= 0; --i) {
+      if (this.recordList[i].driveId === driveInfo.id) {
+        this._recordList.splice(i, 1);
       }
     }
   }
@@ -1653,18 +1660,18 @@ export class FileManagerBase {
     version?: bigint,
     requestOptions?: BeeRequestOptions,
   ): Promise<{ record: FileRecord; fromCache: boolean }> {
-    const cached = this.fileInfoList.find((f) => f.topic === topic);
+    const cached = this.recordList.find((f) => f.topic === topic);
     if (cached) {
       return { record: cached, fromCache: true };
     }
 
     const feedData = await getFeedData(this.bee, new Topic(topic), owner, version, requestOptions);
     if (feedData.feedIndex.equals(FeedIndex.MINUS_ONE)) {
-      throw new FileInfoError(`File record not found for topic: ${topic.slice(0, 6)}`);
+      throw new FileRecordError(`File record not found for topic: ${topic.slice(0, 6)}`);
     }
 
     const loaded = await this.store.getRecord(topic, actPublisher, feedData, requestOptions);
-    this.fileInfoList.push(loaded);
+    this._recordList.push(loaded);
 
     return { record: loaded, fromCache: false };
   }
@@ -1674,14 +1681,14 @@ export class FileManagerBase {
       await this.store.saveRecord(fr, requestOptions);
     } catch (err: unknown) {
       this.errorHandler.handleError(err, `Failed to save record: ${fr.path}`);
-      throw new FileInfoError(`Failed to save record`, err);
+      throw new FileRecordError(`Failed to save record`, err);
     }
 
-    const existingIx = this.fileInfoList.findIndex((f) => f.topic === fr.topic);
+    const existingIx = this.recordList.findIndex((f) => f.topic === fr.topic);
     if (existingIx !== -1) {
-      this.fileInfoList[existingIx] = fr;
+      this._recordList[existingIx] = fr;
     } else {
-      this.fileInfoList.push(fr);
+      this._recordList.push(fr);
     }
   }
 
@@ -1695,10 +1702,10 @@ export class FileManagerBase {
     const isAlreadyTrashed = getRecordStatus(cachedDrive, entry.topic) == NodeStatus.Trashed;
 
     if (isTrashed && isAlreadyTrashed) {
-      throw new FileInfoError(`Already trashed: ${entry.path}`);
+      throw new FileRecordError(`Already trashed: ${entry.path}`);
     }
     if (!isTrashed && !isAlreadyTrashed) {
-      throw new FileInfoError(`Not trashed, cannot recover: ${entry.path}`);
+      throw new FileRecordError(`Not trashed, cannot recover: ${entry.path}`);
     }
 
     const current = cachedDrive.trashedNodes ?? [];
