+++ src/types/index.ts
@@ -1,4 +1,22 @@
-export type { DriveInfo, FileInfo } from './info';
-export { FileStatus } from './info';
-export type { BrowserUploadOptions, FileInfoOptions, NodeUploadOptions, UploadProgress } from './utils';
+export type {
+  DriveInfo,
+  FileRecord,
+  FolderInfo,
+  ManifestHost,
+  NodeEntry,
+  NodeHeader,
+  NodeResource,
+  TrashEntry,
+} from './info';
+export { ListDepth, NodeType, NodeStatus } from './info';
+export type {
+  BrowserUploadOptions,
+  NodeUploadOptions,
+  UploadItem,
+  UpdateItem,
+  UploadSource,
+  UploadFilesResult,
+} from './upload';
+export type { DownloadResource, DownloadResult } from './download';
+export type { ActReferences } from './utils';
 export type { FileManager } from './fileManager';
