+++ src/types/info.ts
@@ -1,36 +1,77 @@
-import { BatchId, EthAddress, FeedIndex, Identifier, PublicKey, RedundancyLevel, Topic } from '@ethersphere/bee-js';
+import { RedundancyLevel } from '@ethersphere/bee-js';
 
-import { ReferenceWithHistory, WrappedFileInfoFeed } from './utils';
+import { ActReferences } from './utils';
 
-// TODO: set statuses for trashed, recovered, forgotten
-export enum FileStatus {
+export enum NodeStatus {
   Active = 'active',
   Trashed = 'trashed',
 }
 
-export interface FileInfo {
-  batchId: string | BatchId;
-  file: ReferenceWithHistory;
-  name: string;
-  owner: string | EthAddress;
-  actPublisher: string | PublicKey;
-  topic: string | Topic;
-  driveId: string;
-  timestamp?: number;
-  preview?: ReferenceWithHistory;
-  version?: string | undefined;
-  index?: FeedIndex | undefined;
-  redundancyLevel?: RedundancyLevel;
-  customMetadata?: Record<string, string>;
-  status?: FileStatus;
+export enum NodeType {
+  File = 'file',
+  Folder = 'folder',
+  Drive = 'drive',
 }
 
-export interface DriveInfo {
-  id: string | Identifier;
-  batchId: string | BatchId;
-  owner: string | EthAddress;
-  name: string;
+export enum ListDepth {
+  Shallow = 'shallow',
+  Deep = 'deep',
+}
+
+export interface NodeResource {
+  batchId: string;
+  topic: string;
+  owner: string;
   redundancyLevel: RedundancyLevel;
+  actPublisher: string;
+  version?: string;
+  status?: NodeStatus;
+}
+
+export interface FileRecord extends NodeResource {
+  type: NodeType.File;
+  driveId: string;
+  path: string;
+  content: ActReferences;
+  timestamp?: number;
+  customMetadata?: Record<string, string>;
+}
+
+export interface ManifestHost extends NodeResource {
+  manifestRef?: ActReferences;
+  version?: never;
+}
+
+export interface TrashEntry {
+  topic: string;
+  type: NodeType;
+  path: string;
+  version?: string;
+}
+
+export interface DriveInfo extends ManifestHost {
+  type: NodeType.Drive;
+  id: string;
+  name: string;
   isAdmin: boolean;
-  infoFeedList?: WrappedFileInfoFeed[];
+  trashedNodes?: TrashEntry[];
+}
+
+export interface FolderInfo extends ManifestHost {
+  type: NodeType.Folder;
+  path: string;
+  driveId: string;
+}
+
+export type NodeEntry = FileRecord | FolderInfo;
+
+export interface NodeHeader {
+  path: string;
+  type: NodeType;
+  topic: string;
+  owner?: string;
+  actPublisher?: string;
+  version?: string;
+  head?: ActReferences;
+  rawMetadata: Record<string, string>;
 }
