--- src/types/v2/index.ts
@@ -1,21 +1,0 @@
-export type {
-  DriveInfo,
-  FileRecord,
-  FolderInfo,
-  ManifestHost,
-  NodeEntry,
-  NodeHeader,
-  NodeResource,
-  TrashEntry,
-} from './info';
-export { ListDepth, NodeType, NodeStatus } from './info';
-export type {
-  BrowserUploadOptions,
-  NodeUploadOptions,
-  UploadItem,
-  UpdateItem,
-  UploadSource,
-  UploadFilesResult,
-} from './upload';
-export type { DownloadResource, DownloadResult } from './download';
-export type { ActReferences, FeedPayloadResult, FeedReferenceResult, FeedResultWithIndex } from './utils';
