--- src/types/v2/info.ts
@@ -1,77 +1,0 @@
-import { RedundancyLevel } from '@ethersphere/bee-js';
-
-import { ActReferences } from './utils';
-
-export enum NodeStatus {
-  Active = 'active',
-  Trashed = 'trashed',
-}
-
-export enum NodeType {
-  File = 'file',
-  Folder = 'folder',
-  Drive = 'drive',
-}
-
-export enum ListDepth {
-  Shallow = 'shallow',
-  Deep = 'deep',
-}
-
-export interface NodeResource {
-  batchId: string;
-  topic: string;
-  owner: string;
-  redundancyLevel: RedundancyLevel;
-  actPublisher: string;
-  version?: string;
-  status?: NodeStatus;
-}
-
-export interface FileRecord extends NodeResource {
-  type: NodeType.File;
-  driveId: string;
-  path: string;
-  content: ActReferences;
-  timestamp?: number;
-  customMetadata?: Record<string, string>;
-}
-
-export interface ManifestHost extends NodeResource {
-  manifestRef?: ActReferences;
-  version?: never;
-}
-
-export interface TrashEntry {
-  topic: string;
-  type: NodeType;
-  path: string;
-  version?: string;
-}
-
-export interface DriveInfo extends ManifestHost {
-  type: NodeType.Drive;
-  id: string;
-  name: string;
-  isAdmin: boolean;
-  trashedNodes?: TrashEntry[];
-}
-
-export interface FolderInfo extends ManifestHost {
-  type: NodeType.Folder;
-  path: string;
-  driveId: string;
-}
-
-export type NodeEntry = FileRecord | FolderInfo;
-
-export interface NodeHeader {
-  path: string;
-  type: NodeType;
-  topic: string;
-  owner?: string;
-  actPublisher?: string;
-  version?: string;
-  head?: ActReferences;
-  rawMetadata: Record<string, string>;
-}
