--- src/utils/v2/index.ts
@@ -1,11 +1,0 @@
-export {
-  assertActReferences,
-  assertDriveInfo,
-  assertDriveInfoFromMetadata,
-  assertFileRecord,
-  assertFolderInfo,
-  assertNodeResource,
-  assertReady,
-  isRecord,
-  parseTrashedNodes,
-} from './asserts';
